requirements.txt
PyYAML==6.0.2
discord==2.3.2
aiohttp==3.11.18
